refactor: Create BaseTypedRepresentationConverter
