+++ src/index.ts
@@ -250,6 +250,7 @@ export * from './storage/accessors/InMemoryDataAccessor';
 export * from './storage/accessors/SparqlDataAccessor';
 
 // Storage/Conversion
+export * from './storage/conversion/BaseTypedRepresentationConverter';
 export * from './storage/conversion/ChainedConverter';
 export * from './storage/conversion/ConstantConverter';
 export * from './storage/conversion/ContainerToTemplateConverter';
