+++ src/storage/conversion/FormToJsonConverter.ts
@@ -5,14 +5,14 @@ import { RepresentationMetadata } from '../../http/representation/Representation
 import { APPLICATION_JSON, APPLICATION_X_WWW_FORM_URLENCODED } from '../../util/ContentTypes';
 import { readableToString } from '../../util/StreamUtil';
 import { CONTENT_TYPE } from '../../util/Vocabularies';
+import { BaseTypedRepresentationConverter } from './BaseTypedRepresentationConverter';
 import type { RepresentationConverterArgs } from './RepresentationConverter';
-import { TypedRepresentationConverter } from './TypedRepresentationConverter';
 
 /**
  * Converts application/x-www-form-urlencoded data to application/json.
  * Due to the nature of form data, the result will be a simple key/value JSON object.
  */
-export class FormToJsonConverter extends TypedRepresentationConverter {
+export class FormToJsonConverter extends BaseTypedRepresentationConverter {
   public constructor() {
     super(APPLICATION_X_WWW_FORM_URLENCODED, APPLICATION_JSON);
   }
