+++ src/storage/conversion/RdfToQuadConverter.ts
@@ -5,13 +5,13 @@ import type { Representation } from '../../http/representation/Representation';
 import { INTERNAL_QUADS } from '../../util/ContentTypes';
 import { BadRequestHttpError } from '../../util/errors/BadRequestHttpError';
 import { pipeSafely } from '../../util/StreamUtil';
+import { BaseTypedRepresentationConverter } from './BaseTypedRepresentationConverter';
 import type { RepresentationConverterArgs } from './RepresentationConverter';
-import { TypedRepresentationConverter } from './TypedRepresentationConverter';
 
 /**
  * Converts most major RDF serializations to `internal/quads`.
  */
-export class RdfToQuadConverter extends TypedRepresentationConverter {
+export class RdfToQuadConverter extends BaseTypedRepresentationConverter {
   public constructor() {
     super(rdfParser.getContentTypesPrioritized(), INTERNAL_QUADS);
   }
